Move callback back to nu-test-support, make diff pretty
