+++ Cargo.toml
@@ -58,7 +58,6 @@ pretty_env_logger = "0.4.0"
 
 [dev-dependencies]
 nu-test-support = { version = "0.29.1", path = "./crates/nu-test-support" }
-nu-plugin-test-support = { version = "0.29.1", path = "./crates/nu-plugin-test-support" }
 dunce = "1.0.1"
 serial_test = "0.5.1"
 hamcrest2 = "0.3.0"
