+++ crates/nu_plugin_inc/Cargo.toml
@@ -16,6 +16,7 @@ nu-protocol = { path = "../nu-protocol", version = "0.29.1" }
 nu-source = { path = "../nu-source", version = "0.29.1" }
 nu-test-support = { path = "../nu-test-support", version = "0.29.1" }
 nu-value-ext = { path = "../nu-value-ext", version = "0.29.1" }
+
 semver = "0.11.0"
 
 [dev-dependencies]
